feat: Comprehensive codebase audit and initial improvements
This commit delivers a comprehensive audit of the codebase and implements several initial improvements based on the findings.

The key contributions of this commit are:

1.  **Codebase Audit:**
    *   A thorough analysis of the application architecture, identifying the client-side data processing as a major scalability bottleneck.
    *   A detailed review of the code quality, which is generally high, with some recommendations for refactoring large components.
    *   An analysis of the application's security and error handling mechanisms.

2.  **UI/UX Enhancements:**
    *   **File Size Validation:** Added a 25MB file size limit for uploads to prevent browser crashes with large files. A user-friendly toast notification is displayed if the limit is exceeded.
    *   **Improved User Feedback:** Implemented toast notifications for creating and updating widgets, providing better feedback for user actions.
    *   **Refactoring for Modularity:** Refactored the `DashboardView.tsx` component by extracting the header into a new, reusable `DashboardHeader.tsx` component. This improves code organization and maintainability.

This work sets the stage for future enhancements, particularly the migration to a full-fledged backend as outlined in the project's planning documents.
